2) remove unnecessary casts to int

+++ get_robust_list.c
@@ -10,7 +10,7 @@
 SYS_FUNC(get_robust_list)
 {
 	if (entering(tcp)) {
-		printpid(tcp, (int) tcp->u_arg[0], PT_TID);
+		printpid(tcp, tcp->u_arg[0], PT_TID);
 		tprints(", ");
 	} else {
 		printnum_ptr(tcp, tcp->u_arg[1]);
